pyproject.toml
[project]
name = "matchbox"
version = "0.1.0"
description = "A framework for orchestrating and comparing various company matching methodologies."
authors = [{ name = "DDaTDataScienceTeam" }]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.4.1",
    "click>=8.1.7",
    "duckdb>=1.1.1",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "pg-bulk-ingest>=0.0.54",
    "psycopg2-binary>=2.9.9",
    "pyarrow>=17.0.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "rustworkx>=0.15.1",
    "splink<4",
    "sqlalchemy>=2.0.35",
    "tomli>=2.0.1",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-env>=1.1.5",
    "ruff>=0.6.8",
    "docker>=7.1.0",
]
package = true

[tool.ruff]
exclude = [
    "*.ipynb"
]
line-length = 88
indent-width = 4
target-version = "py311"
src = ["."]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    # "D"
]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/{test,docs}/*" = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]
addopts = "-s -vv --cov=matchbox test/ --log-disable=pg_bulk_ingest"
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.pytest_env]
MB__POSTGRES__SCHEMA = "test"
MB__BATCH_SIZE = "400"
